+++ src/main/java/nocomment/master/db/TrackResume.java
@@ -4,10 +4,10 @@ import nocomment.master.util.ChunkPos;
 
 public class TrackResume {
     public final ChunkPos pos;
-    public final int dimension;
+    public final short dimension;
     public final long prevTrackID;
 
-    TrackResume(int x, int z, int dimension, long prevTrackID) {
+    TrackResume(int x, int z, short dimension, long prevTrackID) {
         this.pos = new ChunkPos(x, z);
         this.dimension = dimension;
         this.prevTrackID = prevTrackID;
